requirements.txt
pandas==0.24.1
lorem==0.1.1
python-dotenv==0.10.1
gensim==3.4.0
nameparser==1.0.2
spacy==2.0.18
geograpy3==0.1.2
rdflib==4.2.2
SPARQLWrapper==1.8.4